chore: mv tqdm functions to `logging.py`
